fix client import, move checkUserHasMonthlySubscription from universal to server function
